Set up to deploy on Heroku with a Postgres DB
To deploy to Heroku, we'll need to use Postgres in
production. Unfortunately, I've also discovered that if the `sqlite3`
gem is in your `Gemfile`, the installation on Heroku fails. So, let's
switch to using Postgres in development and test too.

It turns out that once you get Postgres running, the setup is actually
quite easy, and you don't need any database passwords in your
`config/database.yml` file.