Fix a stylesheet linking error
The Heroku deployment logs showed that Rails is looking for a
stylesheet called `items.scss.css` and cannot find it.

It turns out that a file whose name ends in `.scss` is transformed
into a `.css` file by the Rails asset pipeline, which runs the SASS
processor on the file. Both the `html.erb` files

    app/views/items/index.html.erb
    app/views/items/show.html.erb

contained a link to this file, written as

    stylesheet_link_tag "items.scss"

THe `stylesheet_link_tag` function turns this into a link to
`/stylesheets/items.scss.css`, which unfortunately doesn't work. For
the link to work, I understand it would have to be

    stylesheet_link_tag "items.css"

or just

    stylesheet_link_tag "items"

However, it turns out that the explicit link to this `.css` file isn't
needed, because of the way the Rails asset pipeline works.

In production, the asset pipeline runs SASS over all the `.css.scss`
files, to produce `.css` files, and then takes all the `.css` files
and concatenates them to make one big `application-...-.css` file
below `public/assets`. All you need is to link to `application.css` in
the 'application.html.erb' layout and you get all the CSS, no matter
which file it was originally in.

In development, it seems to work a little differently, because the
asset pipeline doesn't concatenate all the CSS files to produce one
big CSS file. Instead, it keeps all the files separate, but the link
to `application.css` in 'application.html.erb' expands into a set of
links, one for each CSS file you've got in the application. So again
you don't need to explicitly link to any of the particular CSS files.

Along with removing the explicit links to the `items` CSS file, I've
also removed the `DOCTYPE` declaration and the `<head>` and `<body>`
tags from

    app/views/items/index.html.erb
    app/views/items/show.html.erb

The HTML generated from these files gets 'pasted' into the HTML
generated by `application.html.erb` at the point where it calls
`yield`, so none of those things was needed. The application still
appears the same way, and if you view the source there is now just a
single `<head>` and a single `<body>` section, instead of two of each.

As a precaution, I've added the precompiled assets directory
`public/assets` to .gitignore so that nobody accidentally checks them
in if they run `rake assets:precompile`.